refactor: split poster batch modules
Separates schema, queue, store, task routing, collection, and CLI concerns so the batch workflow can reuse infrastructure pieces across problem types.
